+++ str/str.go
@@ -1,2 +1,0 @@
-// Package str contains string related utilz
-package str