+++ strings/strings.go
@@ -1,0 +1,2 @@
+// Package strings extends the strings package in the stdlib
+package strings