+++ utilz.go
@@ -1,2 +1,2 @@
-// Package utilz contains all kinds of utilities
+// Package utilz contains all kinds of utilities and extensions of the standard library
 package utilz